# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "ASE API Management",
    "description": "The powerful console to manage backend",
    "license": {
      "name": "ASE",
      "url": "http://www.aseit.com.au/"
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
    "/processstoragegrid": {
      "post": {
        "tags": [
          "Storagegrid"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StorageGridInputParams"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New Storagegrid account(tenant account, key creation and bucket creation) has been configured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorageGridSuccessOutputParams"
                }
              }
            }
          },
          "400": {
            "description": "Some issue occured while configuring Storagegrid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorageGridFailureOutputParams"
                }
              }
            }
          },
          "403": {
            "description": "You do not have necessary permissions for the resource"
          }
        }
      }
    },
    "/processveeam": {
      "post": {
        "tags": [
          "Veeam"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VeeamInputParams"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New Veeam account details has been configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VeeamSuccessOutputParams"
                }
              }
            }
          },
          "400": {
            "description": "Some issue occured while configuring Veeam",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VeeamFailureOutputParams"
                }
              }
            }
          },
          "403": {
            "description": "You do not have necessary permissions for the resource"
          }
        }
      }
    },
    "/createnetsuiteclient": {
      "post": {
        "tags": [
          "Netsuite"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NetsuiteCreateClientInputParams"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New client has been configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NetsuiteCreateClientSuccessOutputParams"
                }
              }
            }
          },
          "400": {
            "description": "Some issue occured while configuring new client",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NetsuiteCreateClientFailureOutputParams"
                }
              }
            }
          },
          "403": {
            "description": "You do not have necessary permissions for the resource"
          }
        }
      }
    },
    "/processrequest": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessRequestInputParams"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New details has been configured\n1. Netsuite Client created.\n2. Storagegrid details(tenant account, key creation and bucket creation) has been configured.\n3. Storagegrid account id updated in netsuite\n4. Veeam company has been configured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessRequestSuccessOutputParams"
                }
              }
            }
          },
          "400": {
            "description": "Some issue occured while configuring new details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessRequestFailureOutputParams"
                }
              }
            }
          },
          "403": {
            "description": "You do not have necessary permissions for the resource"
          }
        }
      }
    },
    "/generatetokenforexistingstitchdataaccount": {
      "post": {
        "tags": [
          "Stitchdata"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticateStitchdataInputParams"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stitchdata token has been successfully generated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticateStitchdataSuccessOutputParams"
                }
              }
            }
          },
          "400": {
            "description": "Some issue occured while generating Stitchdata token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessRequestFailureOutputParams"
                }
              }
            }
          },
          "403": {
            "description": "You do not have necessary permissions for the resource"
          }
        }
      }
    },
    "/registeraccountforstitchdata": {
      "post": {
        "tags": [
          "Stitchdata"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterAccountStitchdataInputParams"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stitchdata token has been successfully generated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterAccountForStitchdataSuccessOutputParams"
                }
              }
            }
          },
          "400": {
            "description": "Some issue occured while generating Stitchdata token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessRequestFailureOutputParams"
                }
              }
            }
          },
          "403": {
            "description": "You do not have necessary permissions for the resource"
          }
        }
      }
    },
    "/retrievesourcesfromstitchdata": {
      "post": {
        "tags": [
          "Stitchdata"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetrieveSourcesFromStitchdataInputParams"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stitchdata source has been successfully retrieved.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrieveSourcesFromStitchdataOutputParams"
                }
              }
            }
          },
          "400": {
            "description": "Some issue occured while retrieving sources for Stitchdata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessRequestFailureOutputParams"
                }
              }
            }
          },
          "403": {
            "description": "You do not have necessary permissions for the resource"
          }
        }
      }
    },
    "/retrievedestinationfromstitchdata": {
      "post": {
        "tags": [
          "Stitchdata"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetrieveSourcesFromStitchdataInputParams"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stitchdata destination has been successfully retrieved.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrieveSourcesFromStitchdataOutputParams"
                }
              }
            }
          },
          "400": {
            "description": "Some issue occured while retrieving destination for Stitchdata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessRequestFailureOutputParams"
                }
              }
            }
          },
          "403": {
            "description": "You do not have necessary permissions for the resource"
          }
        }
      }
    },
    "/retrievetenantsforactiveport": {
      "get": {
        "tags": [
          "ActivePort"
        ],
        "responses": {
          "200": {
            "description": "ActivePort token  has been successfully generated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateTokenForActivePortOutputParams"
                }
              }
            }
          },
          "400": {
            "description": "Some issue occured while generating ActivePort token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessRequestFailureOutputParams"
                }
              }
            }
          },
          "403": {
            "description": "You do not have necessary permissions for the resource"
          }
        }
      }
    },
    "/createnewtenantforactiveport": {
      "post": {
        "tags": [
          "ActivePort"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTenantForActivePortInputParams"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ActivePort tenant has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateTokenForActivePortOutputParams"
                }
              }
            }
          },
          "400": {
            "description": "Some issue occured while creating ActivePort tenant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessRequestFailureOutputParams"
                }
              }
            }
          },
          "403": {
            "description": "You do not have necessary permissions for the resource"
          }
        }
      }
    },
    "/updateexistingtenantforactiveport": {
      "put": {
        "tags": [
          "ActivePort"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTenantForActivePortInputParams"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ActivePort tenant has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateTokenForActivePortOutputParams"
                }
              }
            }
          },
          "400": {
            "description": "Some issue occured while creating ActivePort tenant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessRequestFailureOutputParams"
                }
              }
            }
          },
          "403": {
            "description": "You do not have necessary permissions for the resource"
          }
        }
      }
    },
    "/retrieveclustersfromnetapp": {
      "get": {
        "tags": [
          "NetApp"
        ],
        "responses": {
          "200": {
            "description": "NetApp Clusters are retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrieveClustersFromNetappOutputParams"
                }
              }
            }
          },
          "400": {
            "description": "Some issue occured while retrieving NetApp Clusters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessRequestFailureOutputParams"
                }
              }
            }
          },
          "403": {
            "description": "You do not have necessary permissions for the resource"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "StorageGridInputParams": {
        "type": "object",
        "required": [
          "tenantname",
          "bucketname"
        ],
        "properties": {
          "tenantname": {
            "type": "string",
            "uniqueItems": true
          },
          "bucketname": {
            "type": "string",
            "uniqueItems": true
          }
        }
      },
      "VeeamInputParams": {
        "type": "object",
        "required": [
          "name",
          "username",
          "vmsbackedup",
          "vmsbackeduptocloud",
          "managedphysicalservers"
        ],
        "properties": {
          "name": {
            "type": "string",
            "uniqueItems": true
          },
          "username": {
            "type": "string",
            "uniqueItems": true
          },
          "vmsbackedup": {
            "type": "integer",
            "uniqueItems": true
          },
          "vmsbackeduptocloud": {
            "type": "integer",
            "uniqueItems": true
          },
          "managedphysicalservers": {
            "type": "integer",
            "uniqueItems": true
          }
        }
      },
...
      "StorageGridSuccessOutputParams": {
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "properties": {
          "module": {
            "type": "string",
            "uniqueItems": true
          },
          "tenantname": {
            "type": "string",
            "uniqueItems": true
          },
          "tenantpassword": {
            "type": "string",
            "uniqueItems": true
          },
          "bucketname": {
            "type": "string",
            "uniqueItems": true
          },
          "tenantaccountid": {
            "type": "string",
            "uniqueItems": true
          },
          "tenantaccesskey": {
            "type": "string",
            "uniqueItems": true
          },
          "tenantsecretaccesskey": {
            "type": "string",
            "uniqueItems": true
          }
        }
      },
      "VeeamSuccessOutputParams": {
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "properties": {
          "module": {
            "type": "string",
            "uniqueItems": true
          },
          "id": {
            "type": "integer",
            "uniqueItems": true
          },
          "name": {
            "type": "string",
            "uniqueItems": true
          },
          "username": {
            "type": "string",
            "uniqueItems": true
          },
          "password": {
            "type": "string",
            "uniqueItems": true
          }
        }
      },
      "StorageGridFailureOutputParams": {
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "properties": {
          "id": {
            "type": "string",
            "uniqueItems": true
          },
          "logs": {
            "properties": {
              "status": {
                "type": "string",
                "uniqueItems": true
              },
              "description": {
                "type": "string",
                "uniqueItems": true
              },
              "actualdetail": {
                "properties": {
                  "type": "object",
                  "uniqueItems": true
                }
              },
              "param": {
                "properties": {
                  "type": "object",
                  "uniqueItems": true
                }
              },
              "datetime": {
                "type": "string",
                "uniqueItems": true
              }
            }
          }
        }
      },
      "VeeamFailureOutputParams": {
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "properties": {
          "id": {
            "type": "string",
            "uniqueItems": true
          },
          "logs": {
            "properties": {
              "status": {
                "type": "string",
                "uniqueItems": true
              },
              "description": {
                "type": "string",
                "uniqueItems": true
              },
              "actualdetail": {
                "properties": {
                  "type": "object",
                  "uniqueItems": true
                }
              },
              "param": {
                "properties": {
                  "type": "object",
                  "uniqueItems": true
                }
              },
              "datetime": {
                "type": "string",
                "uniqueItems": true
              }
            }
          }
        }
      },
      "NetsuiteCreateClientInputParams": {
        "type": "object",
        "required": [
          "clientname",
          "acn"
        ],
        "properties": {
          "clientname": {
            "type": "string",
            "uniqueItems": true
          },
          "acn": {
            "type": "string",
            "uniqueItems": true
          }
        }
      },
      "NetsuiteCreateClientSuccessOutputParams": {
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "properties": {
          "module": {
            "type": "string",
            "uniqueItems": true
          },
          "entityId": {
            "type": "string",
            "uniqueItems": true
          },
          "clientid": {
            "type": "string",
            "uniqueItems": true
          },
          "clientName": {
            "type": "string",
            "uniqueItems": true
          },
          "ACN": {
            "type": "string",
            "uniqueItems": true
          }
        }
      },
      "NetsuiteCreateClientFailureOutputParams": {
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "properties": {
          "id": {
            "type": "string",
            "uniqueItems": true
          },
          "logs": {
            "properties": {
              "status": {
                "type": "string",
                "uniqueItems": true
              },
              "description": {
                "type": "string",
                "uniqueItems": true
              },
              "actualdetail": {
                "properties": {
                  "type": "object",
                  "uniqueItems": true
                }
              },
              "param": {
                "properties": {
                  "type": "object",
                  "uniqueItems": true
                }
              },
              "datetime": {
                "type": "string",
                "uniqueItems": true
              }
            }
          }
        }
      },
      "ProcessRequestInputParams": {
        "type": "object",
        "required": [
          "bucketname",
          "clientname",
          "acn",
          "address",
          "vmsbackedup",
          "vmsbackeduptocloud",
          "managedphysicalservers"
        ],
        "properties": {
          "bucketname": {
            "type": "string",
            "uniqueItems": true
          },
          "clientname": {
            "type": "string",
            "uniqueItems": true
          },
          "acn": {
            "type": "string",
            "uniqueItems": true
          },
          "address": {
            "type": "string",
            "uniqueItems": true
          },
          "vmsbackedup": {
            "type": "integer",
            "uniqueItems": true
          },
          "vmsbackeduptocloud": {
            "type": "integer",
            "uniqueItems": true
          },
          "managedphysicalservers": {
            "type": "integer",
            "uniqueItems": true
          }
        }
      },
      "ProcessRequestSuccessOutputParams": {
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "properties": {
          "NetSuiteResponse": {
            "$ref": "#/components/schemas/NetsuiteCreateClientSuccessOutputParams"
          },
          "StorageGridResponse": {
            "$ref": "#/components/schemas/StorageGridSuccessOutputParams"
          },
          "VeeamResponse": {
            "$ref": "#/components/schemas/VeeamSuccessOutputParams"
          }
        }
      },
      "ProcessRequestFailureOutputParams": {
        "type": "array",
        "required": [
          "success",
          "message"
        ],
        "items": {
          "$ref": "#/components/schemas/NetsuiteCreateClientFailureOutputParams"
        }
      },
      "AuthenticateStitchdataInputParams": {
        "type": "object",
        "required": [
          "stitch_auth_code"
        ],
        "properties": {
          "stitch_auth_code": {
            "type": "string",
            "uniqueItems": true
          }
        }
      },
      "AuthenticateStitchdataSuccessOutputParams": {
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "properties": {
          "module": {
            "type": "string",
            "uniqueItems": true
          },
          "access_token": {
            "type": "string",
            "uniqueItems": true
          },
          "token_type": {
            "type": "string",
            "uniqueItems": true
          },
          "stitch_account_id": {
            "type": "string",
            "uniqueItems": true
          }
        }
      },
      "RegisterAccountForStitchdataSuccessOutputParams": {
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "properties": {
          "module": {
            "type": "string",
            "uniqueItems": true
          },
          "firstname": {
            "type": "string",
            "uniqueItems": true
          },
          "lastname": {
            "type": "string",
            "uniqueItems": true
          },
          "company": {
            "type": "string",
            "uniqueItems": true
          },
          "email": {
            "type": "string",
            "uniqueItems": true
          },
          "access_token": {
            "type": "string",
            "uniqueItems": true
          },
          "stitch_account_id": {
            "type": "string",
            "uniqueItems": true
          }
        }
      },
...
      "RegisterAccountStitchdataInputParams": {
        "type": "object",
        "required": [
          "firstname",
          "lastname",
          "company",
          "email"
        ],
        "properties": {
          "firstname": {
            "type": "string",
            "uniqueItems": true
          },
          "lastname": {
            "type": "string",
            "uniqueItems": true
          },
          "company": {
            "type": "string",
            "uniqueItems": true
          },
          "email": {
            "type": "string",
            "uniqueItems": true
          }
        }
      },
      "RetrieveSourcesFromStitchdataInputParams": {
        "type": "object",
        "required": [
          "access_token"
        ],
        "properties": {
          "access_token": {
            "type": "string",
            "uniqueItems": true
          }
        }
      },
      "RetrieveSourcesFromStitchdataOutputParams": {
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "properties": {
          "module": {
            "type": "string",
            "uniqueItems": true
          },
          "access_token": {
            "type": "string",
            "uniqueItems": true
          },
          "message": {
            "type": "object",
            "uniqueItems": true
          }
        }
      },
...
      "CreateTenantForActivePortInputParams": {
        "type": "object",
        "required": [
          "id",
          "tenantId",
          "description",
          "name",
          "servicesList"
        ],
        "properties": {
          "description": {
            "type": "string",
            "uniqueItems": true
          },
          "name": {
            "type": "string",
            "uniqueItems": true
          },
          "servicesList": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          }
        }
      },
...
      "UpdateTenantForActivePortInputParams": {
        "type": "object",
        "required": [
          "id",
          "tenantId",
          "description",
          "name",
          "servicesList"
        ],
        "properties": {
          "id": {
            "type": "string",
            "uniqueItems": true
          },
          "tenantId": {
            "type": "string",
            "uniqueItems": true
          },
          "description": {
            "type": "string",
            "uniqueItems": true
          },
          "name": {
            "type": "string",
            "uniqueItems": true
          },
          "servicesList": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          }
        }
      },
...
    },
...
  },
...
}